Show All Option Details. Fix Double Message Bug
When a user responded yes to an option, the logic would send them a success message
when the status of the Plan updated, but ALSO would notice they haven't been asked
about the winning option with a single_option answer. We would then ask them if
they wanted to join with the winning option. This logic was supposed to accommodate
late responders, and allow them to join a plan that had already been decided. Instead
we're moving that logic next to the bit where we update their availability.