Late Comer Bugfix, Text Changes, Self-Channel Bugfix
Changed some text to make it a bit longer, more fun, clearer.
There was an issue when you tried to start /letshang from your own
self-chat channel. The fix is to treat a DM differently.
When everyone agreed to a plan, and we asked folks  that hadn't
originally said they wanted to do that option, that user would
never get the message that everyone else had agreed to do something.